requirements.txt
pillow>=11.0
pygame-ce>=2.5.7
pytmx>=3.32
repodnet>=0.1.2
